minor modifications and services unit testing added

+++ src/main/java/com/ChaoticChaotic/db2/services/impl/ShippingsImpl.java
@@ -1,12 +1,13 @@
 package com.ChaoticChaotic.db2.services.impl;
 
 import com.ChaoticChaotic.db2.entity.Shippings;
+import com.ChaoticChaotic.db2.exception.BadRequestException;
+import com.ChaoticChaotic.db2.exception.IdNotFoundException;
 import com.ChaoticChaotic.db2.repository.ShippingsRepository;
 import com.ChaoticChaotic.db2.services.ShippingsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -16,23 +17,31 @@ public class ShippingsImpl implements ShippingsService {
     private ShippingsRepository shippingsRepository;
 
 
-    public String addShipping(Shippings shipping) {
-    if (shipping.getStartDate().isBefore(shipping.getEndDate())){
-       shippingsRepository.save(shipping);
+    public void addShipping(Shippings shipping) {
+    if (!shipping.getStartDate().isBefore(shipping.getEndDate())){
+            throw new BadRequestException(
+                    "Incorrect shipping dates!"
+            );
         }
-        return "Saved";
+        shippingsRepository.save(shipping);
     }
 
 
-    public String deleteShipping(Long id) {
-    shippingsRepository.deleteById(id);
-        return "Deleted";
+    public void deleteShipping(Long id) {
+        if(!shippingsRepository.existsById(id)) {
+            throw new IdNotFoundException(
+                    "Line with id " + id + " does not exists");
+        }
+        shippingsRepository.deleteById(id);
     }
 
 
     public List<Shippings> showShipping() {
-        List<Shippings> allShippings = new ArrayList<>();
-        shippingsRepository.findAll().forEach(shipping -> allShippings.add(shipping));
-        return allShippings;
+        return shippingsRepository.findAll();
     }
+
+    public ShippingsImpl(ShippingsRepository shippingsRepository) {
+        this.shippingsRepository = shippingsRepository;
+    }
+
 }
